requirements.txt
kafka-python
flask
snappy